refactored StoryNodeModal, split it into Show + Edit
